# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FluxScale Control API",
    "version": "0.5.0-rc.1",
    "description": "Single-controller Docker autoscaling. Ingest authorization depends on execution mode; read access is a distinct role. Replica reconciliation is asynchronous. Public health follows configuration."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/services/{service}/instances": {
      "parameters": [
        {
          "$ref": "#/components/parameters/Service"
        }
      ],
      "get": {
        "summary": "Read-role per-instance samples and freshness",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Json"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/dashboard/overview": {
      "get": {
        "summary": "Read-role fleet overview, estimated capacity, decisions and persistence",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Json"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/audit": {
      "get": {
        "summary": "Read-role bounded credential-free request audit",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Json"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Distinct read, ingest and managed tokens. Exact route/execution-mode role required. No token values belong in URLs."
      }
    },
    "parameters": {
      "Service": {
        "name": "service",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Json": {
        "description": "JSON response; every response carries X-Request-Id",
        "content": {
          "application/json": {
            "schema": {}
          }
        }
      },
      "Error": {
        "description": "JSON error; 401 unauthenticated, 403 wrong role, 404 unknown service, 429 rate limited. Rejected ingestion does not mutate service state.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
...
    },
...
  }
}